change userId to number,change auth middleware
