+++ .env.dist
@@ -6,3 +6,4 @@ DB_HOST=localhost
 NODE_ENV=development
 SENDGRID_API_KEY=
 SENDGRID_EMAIL_SENDER=
+JWT_SECRET=
